requirements.txt
kafka-python>=2.0.2
apache-flink>=1.17.0
apache-flink-libraries>=1.17.0
python-dateutil>=2.8.2
